refactor: Cleaner utils structure
