+++ tests/pipelines/pipeline/test_s3_folder_received.py
@@ -16,8 +16,8 @@ from dpypelines.pipeline.validate_pipeline import retrieve_config_and_files
 from dpypelines.s3_folder_received import start
 
 
-@patch("dpypelines.pipeline.utils.get_notifier")
-@patch("dpypelines.pipeline.utils.get_email_client")
+@patch("dpypelines.pipeline.messages.utils.get_notifier")
+@patch("dpypelines.pipeline.messages.utils.get_email_client")
 def test_setup_clients(mock_get_email_client, mock_get_notifier):
     """Test that `setup_clients()` returns the expected clients."""
     mock_notifier = MagicMock()
